phase4/requirements.txt
# =============================================================================
# AgentMeshKafka - Python Dependencies (Phase 4)
# =============================================================================
# Install: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Core LLM Framework
# -----------------------------------------------------------------------------
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# -----------------------------------------------------------------------------
# Anthropic Claude SDK (Primary LLM - see docs/07-Constitution.md)
# -----------------------------------------------------------------------------
anthropic>=0.40.0

# -----------------------------------------------------------------------------
# Apache Kafka & Schema Registry
# -----------------------------------------------------------------------------
confluent-kafka>=2.3.0
fastavro>=1.9.0

# -----------------------------------------------------------------------------
# RAG & Vector Store
# -----------------------------------------------------------------------------
chromadb>=0.4.22
sentence-transformers>=2.2.0

# -----------------------------------------------------------------------------
# Data Validation & Serialization
# -----------------------------------------------------------------------------
pydantic>=2.5.0
pydantic-settings>=2.1.0

# -----------------------------------------------------------------------------
# Async & HTTP
# -----------------------------------------------------------------------------
httpx>=0.25.0
aiohttp>=3.9.0

# -----------------------------------------------------------------------------
# Environment & Configuration
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# Observability & Logging
# -----------------------------------------------------------------------------
structlog>=24.1.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-logging>=0.43b0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-instrumentation-httpx>=0.43b0

# -----------------------------------------------------------------------------
# Testing & Evaluation (AgentOps)
# -----------------------------------------------------------------------------
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
deepeval>=0.21.0

# -----------------------------------------------------------------------------
# Development Tools
# -----------------------------------------------------------------------------
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
pre-commit>=3.6.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
rich>=13.7.0
typer>=0.9.0
